feat: Split Monitor client into EventsClient clients
