+++ sdcclient/_common.py
@@ -1,5 +1,6 @@
-import os
 import json
+import os
+
 import requests
 
 
@@ -37,7 +38,7 @@ class _SdcCommon(object):
         return headers
 
     def _checkResponse(self, res):
-        if res.status_code >= 300:
+        if res.status_code >= 300:  # FIXME: Should it be >=400? 301 = Moved Permanently, 302 = Found, 303 = See Other
             errorcode = res.status_code
             self.lasterr = None
 
@@ -359,112 +360,6 @@ class _SdcCommon(object):
         res = requests.post(self.url + '/api/data?format=map', headers=self.hdrs,
                             data=json.dumps(req_json), verify=self.ssl_verify)
         return self._request_result(res)
-
-    def post_event(self, name, description=None, severity=None, event_filter=None, tags=None):
-        '''**Description**
-            Send an event to Sysdig Monitor. The events you post are available in the Events tab in the Sysdig Monitor UI and can be overlied to charts.
-
-        **Arguments**
-            - **name**: the name of the new event.
-            - **description**: a longer description offering detailed information about the event.
-            - **severity**: syslog style from 0 (high) to 7 (low).
-            - **event_filter**: metadata, in Sysdig Monitor format, of nodes to associate with the event, e.g. ``host.hostName = 'ip-10-1-1-1' and container.name = 'foo'``.
-            - **tags**: a list of key-value dictionaries that can be used to tag the event. Can be used for filtering/segmenting purposes in Sysdig Monitor.
-
-        **Success Return Value**
-            A dictionary describing the new event.
-
-        **Examples**
-            - `examples/post_event_simple.py <https://github.com/draios/python-sdc-client/blob/master/examples/post_event_simple.py>`_
-            - `examples/post_event.py <https://github.com/draios/python-sdc-client/blob/master/examples/post_event.py>`_
-        '''
-        options = {
-            'name': name,
-            'description': description,
-            'severity': severity,
-            'filter': event_filter,
-            'tags': tags
-        }
-        edata = {
-            'event': {k: v for k, v in options.items() if v is not None}
-        }
-        res = requests.post(self.url + '/api/events/', headers=self.hdrs, data=json.dumps(edata),
-                            verify=self.ssl_verify)
-        return self._request_result(res)
-
-    def get_events(self, name=None, category=None, direction='before', status=None, limit=100, pivot=None):
-        '''**Description**
-            Returns the list of Sysdig Monitor events.
-
-        **Arguments**
-            - **name**: filter events by name. Default: None.
-            - **category**: filter events by category. Default: ['alert', 'custom', 'docker', 'containerd', 'kubernetes'].
-            - **direction**: orders the list of events. Valid values: "before", "after". Default: "before".
-            - **status**: status of the event as list. Default: ['triggered', 'resolved', 'acknowledged', 'unacknowledged']
-            - **limit**: max number of events to retrieve. Default: 100.
-            - **pivot**: event id to use as pivot. Default: None.
-
-        **Success Return Value**
-            A dictionary containing the list of events.
-
-        **Example**
-            `examples/list_events.py <https://github.com/draios/python-sdc-client/blob/master/examples/list_events.py>`_
-        '''
-        valid_categories = ['alert', 'custom', 'docker', 'containerd', 'kubernetes']
-
-        if category is None:
-            category = valid_categories
-
-        for c in category:
-            if c not in valid_categories:
-                return False, "Invalid category '{}'".format(c)
-
-        valid_status = ["triggered", "resolved", "acknowledged", "unacknowledged"]
-        if status is None:
-            status = valid_status
-
-        for s in status:
-            if s not in valid_status:
-                return False, "Invalid status '{}'".format(s)
-
-        if direction not in ["before", "after"]:
-            return False, "Invalid direction '{}', must be either 'before' or 'after'".format(direction)
-
-        options = {
-            'alertStatus': status,
-            'category': ','.join(category),
-            'dir': direction,
-            'feed': 'true',
-            'include_pivot': 'true',
-            'include_total': 'true',
-            'limit': str(limit),
-            'pivot': pivot,
-            'filter': name,
-        }
-        params = {k: v for k, v in options.items() if v is not None}
-        res = requests.get(self.url + '/api/v2/events/', headers=self.hdrs, params=params, verify=self.ssl_verify)
-        return self._request_result(res)
-
-    def delete_event(self, event):
-        '''**Description**
-            Deletes an event.
-
-        **Arguments**
-            - **event**: the event object as returned by :func:`~SdcClient.get_events`.
-
-        **Success Return Value**
-            `None`.
-
-        **Example**
-            `examples/delete_event.py <https://github.com/draios/python-sdc-client/blob/master/examples/delete_event.py>`_
-        '''
-        if 'id' not in event:
-            return [False, "Invalid event format"]
-
-        res = requests.delete(self.url + '/api/events/' + str(event['id']), headers=self.hdrs, verify=self.ssl_verify)
-        if not self._checkResponse(res):
-            return [False, self.lasterr]
-        return [True, None]
 
     def get_data(self, metrics, start_ts, end_ts=0, sampling_s=0,
                  filter='', datasource_type='host', paging=None):
@@ -1061,7 +956,8 @@ class _SdcCommon(object):
         **Reslut**
             The access keys object
         '''
-        res = requests.post(self.url + '/api/customer/accessKeys/' + access_key + "/disable/", headers=self.hdrs, verify=self.ssl_verify)
+        res = requests.post(self.url + '/api/customer/accessKeys/' + access_key + "/disable/", headers=self.hdrs,
+                            verify=self.ssl_verify)
         return self._request_result(res)
 
     def enable_access_key(self, access_key):
@@ -1075,7 +971,8 @@ class _SdcCommon(object):
         **Reslut**
             The access keys object
         '''
-        res = requests.post(self.url + '/api/customer/accessKeys/' + access_key + "/enable/", headers=self.hdrs, verify=self.ssl_verify)
+        res = requests.post(self.url + '/api/customer/accessKeys/' + access_key + "/enable/", headers=self.hdrs,
+                            verify=self.ssl_verify)
         return self._request_result(res)
 
     def get_agents_config(self):
