+++ sdcclient/monitor/__init__.py
@@ -1,0 +1,2 @@
+from ._events_v1 import EventsClientV1
+from ._events_v2 import EventsClientV2
